requirements.txt
slack-bolt==1.15.2
python-dotenv==0.21.0
openai==0.27.8
tiktoken==0.3.3
pillow
youtube-transcript-api
extractcontent3
gradio_client
